+++ examples/scripts/spm_pso.py
@@ -37,7 +37,7 @@ dataset = pybop.Dataset(
 # Generate problem, cost function, and optimisation class
 problem = pybop.FittingProblem(model, parameters, dataset)
 cost = pybop.SumSquaredError(problem)
-optim = pybop.PSO(cost, max_iterations=100)
+optim = pybop.Optimisation(cost, optimiser=pybop.PSO, max_iterations=100)
 
 x, final_cost = optim.run()
 print("Estimated parameters:", x)