+++ tests/integration/test_optimisation_options.py
@@ -80,8 +80,9 @@ class TestOptimisation:
     )
     @pytest.mark.integration
     def test_optimisation_f_guessed(self, f_guessed, spm_costs):
+        x0 = spm_costs.x0
         # Test each optimiser
-        parameterisation = pybop.XNES(
+        optim = pybop.XNES(
             cost=spm_costs,
             sigma0=0.05,
             max_iterations=125,
@@ -92,16 +93,17 @@ class TestOptimisation:
 
         # Set parallelisation if not on Windows
         if sys.platform != "win32":
-            parameterisation.set_parallel(True)
+            optim.set_parallel(True)
 
-        initial_cost = parameterisation.cost(spm_costs.x0)
-        x, final_cost = parameterisation.run()
+        initial_cost = optim.cost(x0)
+        x, final_cost = optim.run()
 
         # Assertions
-        if parameterisation.minimising:
-            assert initial_cost > final_cost
-        else:
-            assert initial_cost < final_cost
+        if not np.allclose(x0, self.ground_truth, atol=1e-5):
+            if optim.minimising:
+                assert initial_cost > final_cost
+            else:
+                assert initial_cost < final_cost
         np.testing.assert_allclose(x, self.ground_truth, atol=2.5e-2)
 
     def getdata(self, model, x, init_soc):