+++ tests/unit/test_optimisation.py
@@ -90,6 +90,12 @@ class TestOptimisation:
         assert optim.cost is not None
         assert optim.name() == expected_name
 
+        # Test pybop.Optimisation construction
+        optim = pybop.Optimisation(cost=cost, optimiser=optimiser)
+
+        assert optim.cost is not None
+        assert optim.name() == expected_name
+
         if optimiser not in [pybop.SciPyDifferentialEvolution]:
             # Test construction without bounds
             optim = optimiser(cost=cost, bounds=None)
@@ -235,7 +241,7 @@ class TestOptimisation:
             Exception,
             match="The cost is not a recognised cost object or function.",
         ):
-            pybop.DefaultOptimiser(cost="Invalid string")
+            pybop.Optimisation(cost="Invalid string")
 
         def invalid_cost(x):
             return [1, 2]
@@ -244,12 +250,19 @@ class TestOptimisation:
             Exception,
             match="not a scalar numeric value.",
         ):
-            pybop.DefaultOptimiser(cost=invalid_cost)
+            pybop.Optimisation(cost=invalid_cost)
 
     @pytest.mark.unit
     def test_default_optimiser(self, cost):
-        optim = pybop.DefaultOptimiser(cost=cost)
+        optim = pybop.Optimisation(cost=cost)
         assert optim.name() == "Exponential Natural Evolution Strategy (xNES)"
+
+        # Test incorrect setting attribute
+        with pytest.raises(
+            AttributeError,
+            match="'Optimisation' object has no attribute 'not_a_valid_attribute'",
+        ):
+            optim.not_a_valid_attribute
 
     @pytest.mark.unit
     def test_incorrect_optimiser_class(self, cost):
@@ -263,13 +276,16 @@ class TestOptimisation:
             pybop.BasePintsOptimiser(cost=cost, pints_optimiser=RandomClass)
 
         with pytest.raises(NotImplementedError):
-            pybop.Optimisation(cost=cost)
+            pybop.BaseOptimiser(cost=cost)
+
+        with pytest.raises(ValueError):
+            pybop.Optimisation(cost=cost, optimiser=RandomClass)
 
     @pytest.mark.unit
     def test_prior_sampling(self, cost):
         # Tests prior sampling
         for i in range(50):
-            optim = pybop.DefaultOptimiser(cost=cost)
+            optim = pybop.Optimisation(cost=cost)
 
             assert optim.x0 <= 0.62 and optim.x0 >= 0.58
 
@@ -342,7 +358,7 @@ class TestOptimisation:
         with pytest.raises(ValueError):
             optim.set_max_unchanged_iterations(1, threshold=-1)
 
-        optim = pybop.DefaultOptimiser(cost=cost)
+        optim = pybop.Optimisation(cost=cost)
 
         # Trigger threshold
         optim._threshold = np.inf
@@ -387,5 +403,5 @@ class TestOptimisation:
     @pytest.mark.unit
     def test_unphysical_result(self, cost):
         # Trigger parameters not physically viable warning
-        optim = pybop.DefaultOptimiser(cost=cost)
+        optim = pybop.Optimisation(cost=cost)
         optim.check_optimal_parameters(np.array([2]))